explicit destroy() is faster than weaken
Calling Scalar::Util::weaken has some overhead:
about 1.5s (5% of the total benchmark time).
When we want to free an object (document with all its nodes),
weaken-version takes 1.3s and destroy-version takes 1.1s.
Sub destroy() explicitly breaks the cyclic references.
When Perl exits, weaken-version is 0.8s faster
because most of the garbage collection was done already
when the $doc went out of the scope.
However, the exit time is not important in most cases
because all the documents are already processed/saved.

So without weaken we can save even more than the 1.5s,
but when processing more documents,
we must use $doc->destroy() to prevent memory leaks.
Even this could be automatized using something like Object::Destroyer,
but it is not worth the work because Treex::Core::Scenario
will call $doc->destroy(); (so most users do not need to know about it).