without $LASTCHILD pointer
The $LASTCHILD pointer is not needed if the linked list is not sorted.

Even without this pointer the bi-linked list of children does not bring any speedup.
It makes "rehang" a bit faster, but "add" and "load" slower.
The reason is that on average a node has 0.9 children
and non-leaf node 2.3 - 2.5 children.
So on average the linked lists are not long
and iterating the list during "rehang" is fast.