add Write::Treex
It may be useful if you want to visualize the documents in TrEd (well, ttred).
If you have the newest Treex installed, it should be better
to convert the file from conllu to treex.gz via Treex:
 treex Read::CoNLLU from=my.conllu Write::Treex
 ttred my.treex.gz
But if you have just the EasyTreex extension of TrEd,
you may find this block useful.